+++ Documents/Help/Changelog.md
@@ -1,4 +1,13 @@
-# <img src="https://github.com/Krypton-Suite/Standard-Toolkit-Demos/blob/master/Krypton.png"> Standard Toolkit Demos - Changelog
+# <img src="https://github.com/Krypton-Suite/Standard-Toolkit-Demos/blob/master/Krypton.png?raw=true"> Standard Toolkit Demos - ChangeLog
+
+## 2022-12-xx - Build 2212 - December 2022
+* Resolves [#112](https://github.com/Krypton-Suite/Standard-Toolkit-Demos/issues/112), Alpha Branch does not build against "Alpha" of the Toolkit
+* Fixed theme bug for `KryptonExplorer`
+
+## 2022-11-21 - Build 2211 - November 2022
+* Add support for .NET 7
+* Removed support for .NET 5
+* API tweaks
 
 ## 2022-01-05 - Build 2201 - January 2022
 * Added double click functionality to combo box, list box, list view, tree view etc demos